Extracted numimages, sorted imagemode
